+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Version 1.0.9
 -------------
 * Align with Java version numbering
 * Add parsing of independent padding attributes
+* pstsTextAlpha value range change from 0..1 to 0..255
+* New attr pstsTextColorSelected
 
 
 Version 1.0.1